+++ src/main/java/com/project/Project/service/impl/FavoriteServiceImpl.java
@@ -3,20 +3,27 @@ package com.project.Project.service.impl;
 import com.project.Project.domain.Member;
 import com.project.Project.domain.building.Building;
 import com.project.Project.domain.interaction.Favorite;
+import com.project.Project.repository.building.BuildingCustomRepository;
 import com.project.Project.repository.building.BuildingRepository;
 import com.project.Project.repository.interaction.FavoriteRepository;
 import com.project.Project.service.FavoriteService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class FavoriteServiceImpl implements FavoriteService {
 
+
     private final BuildingRepository buildingRepository;
     private final FavoriteRepository favoriteRepository;
+    private final BuildingCustomRepository buildingCustomRepo;
 
     @Transactional
     public Long deleteFavoriteBuilding(Long buildingId, Member member) {
@@ -34,5 +41,13 @@ public class FavoriteServiceImpl implements FavoriteService {
         favorite.setMember(member);
         Favorite savedFavorite = favoriteRepository.save(favorite);
         return savedFavorite.getId();
+    }
+
+    public List<Building> getBuildingListByMember(Member member, Long cursorId, Pageable pageable) {
+        List<Favorite> favoriteList = favoriteRepository.findByMember(member);
+        List<Long> buildingIds = favoriteList.stream()
+                .map(favorite -> favorite.getBuilding().getId())
+                .collect(Collectors.toList());
+        return buildingCustomRepo.findBuildingsByIdIn(buildingIds, cursorId, pageable);
     }
 }
